requirements.txt
flask
groq
langchain-community
sentence-transformers
faiss-cpu
python-dotenv
pypdf
langchain-text-splitters

